Id类型改为long.

+++ Source/Starfish.Service/Application/Commands/Team/TeamUpdateCommand.cs
@@ -5,13 +5,13 @@ namespace Nerosoft.Starfish.Application;
 
 public class TeamUpdateCommand : Command
 {
-	public TeamUpdateCommand(int id, TeamEditDto data)
+	public TeamUpdateCommand(long id, TeamEditDto data)
 	{
 		Id = id;
 		Data = data;
 	}
 
-	public int Id { get; set; }
+	public long Id { get; set; }
 
 	public TeamEditDto Data { get; set; }
 }
