+++ Source/Starfish.Service/Application/Contracts/IUserApplicationService.cs
@@ -14,7 +14,7 @@ public interface IUserApplicationService : IApplicationService
 	/// <param name="data"></param>
 	/// <param name="cancellationToken"></param>
 	/// <returns></returns>
-	Task<int> CreateAsync(UserCreateDto data, CancellationToken cancellationToken = default);
+	Task<long> CreateAsync(UserCreateDto data, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// 编辑用户
@@ -23,7 +23,7 @@ public interface IUserApplicationService : IApplicationService
 	/// <param name="data"></param>
 	/// <param name="cancellationToken"></param>
 	/// <returns></returns>
-	Task UpdateAsync(int id, UserUpdateDto data, CancellationToken cancellationToken = default);
+	Task UpdateAsync(long id, UserUpdateDto data, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// 查询符合条件的用户列表
@@ -49,7 +49,7 @@ public interface IUserApplicationService : IApplicationService
 	/// <param name="id"></param>
 	/// <param name="cancellationToken"></param>
 	/// <returns></returns>
-	Task<UserDetailDto> GetAsync(int id, CancellationToken cancellationToken = default);
+	Task<UserDetailDto> GetAsync(long id, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// 删除用户
@@ -57,7 +57,7 @@ public interface IUserApplicationService : IApplicationService
 	/// <param name="id"></param>
 	/// <param name="cancellationToken"></param>
 	/// <returns></returns>
-	Task DeleteAsync(int id, CancellationToken cancellationToken = default);
+	Task DeleteAsync(long id, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// 设置用户角色
@@ -66,7 +66,7 @@ public interface IUserApplicationService : IApplicationService
 	/// <param name="roles"></param>
 	/// <param name="cancellationToken"></param>
 	/// <returns></returns>
-	Task SetRolesAsync(int id, List<string> roles, CancellationToken cancellationToken = default);
+	Task SetRolesAsync(long id, List<string> roles, CancellationToken cancellationToken = default);
 
 	/// <summary>
 	/// 修改密码
@@ -84,5 +84,5 @@ public interface IUserApplicationService : IApplicationService
 	/// <param name="password"></param>
 	/// <param name="cancellationToken"></param>
 	/// <returns></returns>
-	Task ResetPasswordAsync(int id, string password, CancellationToken cancellationToken = default);
+	Task ResetPasswordAsync(long id, string password, CancellationToken cancellationToken = default);
 }
