+++ Source/Starfish.Service/UseCases/Identity/UserDeleteUseCase.cs
@@ -6,7 +6,7 @@ namespace Nerosoft.Starfish.UseCases;
 
 public interface IUserDeleteUseCase : INonOutputUseCase<UserDeleteInput>;
 
-public record UserDeleteInput(int Id) : IUseCaseInput;
+public record UserDeleteInput(long Id) : IUseCaseInput;
 
 public class UserDeleteUseCase : IUserDeleteUseCase
 {
