+++ Source/Starfish.Transit/Team/TeamItemDto.cs
@@ -8,7 +8,7 @@ public class TeamItemDto
 	/// <summary>
 	/// Id
 	/// </summary>
-	public int Id { get; set; }
+	public long Id { get; set; }
 
 	/// <summary>
 	/// 别名
@@ -28,7 +28,7 @@ public class TeamItemDto
 	/// <summary>
 	/// 团队所有者Id
 	/// </summary>
-	public int OwnerId { get; set; }
+	public long OwnerId { get; set; }
 
 	/// <summary>
 	/// 成员数量
